renamed beans -> domain
