+++ link-shortener/src/main/java/com/google/cloud/servicebroker/examples/linkshortener/domain/LinkInfo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.servicebroker.examples.linkshortener.beans;
+package com.google.cloud.servicebroker.examples.linkshortener.domain;
 
 import com.google.cloud.servicebroker.examples.linkshortener.enums.ThreatStatus;
