+++ link-shortener/src/main/java/com/google/cloud/servicebroker/examples/linkshortener/repositories/LinkRepository.java
@@ -15,7 +15,7 @@
  */
 package com.google.cloud.servicebroker.examples.linkshortener.repositories;
 
-import com.google.cloud.servicebroker.examples.linkshortener.beans.Link;
+import com.google.cloud.servicebroker.examples.linkshortener.domain.Link;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
